Factor out bookable-item and prettify
