+++ .gitignore
@@ -7,3 +7,4 @@ expo-env.d.ts
 
 node_modules
 .expo
+dist
